Player Work: Movement Cost Indicator & Cleanup
- Added a text box that shows the cost of a move above the currently selected tile
- - - Straight moves cost 1, diagonal moves cost 1.5
- - - This added cost, including halves, is shown to the player
- ExtraTurns on player is now tracked as a float
- - - This was all done because players could get around the previous move limit system by moving one diagonal at a time. Allowing them to move 4 tiles diagonally when 3 should be the max. This balancing seemed unclear without a move cost being given to the player.

- Removed unused code. Some variables are now fetched on Awake() instead of in inspector

+++ Assets/Scripts/Entities/AbstractEntity.cs
@@ -14,7 +14,7 @@ namespace Entities
         public bool fourDir;
         private int lostTurns;
         private int hitPoints = 1;
-        protected int extraTurns;
+        protected float extraTurns;
         private const int MAX_HEALTH = 2;
         protected Action currentFinishAction;
 
@@ -145,7 +145,7 @@ namespace Entities
             extraTurns = range;
         }
 
-        public int GetTurns()
+        public float GetTurns()
         {
             return extraTurns;
         }
